pyproject.toml
[tool.poetry]
name = "ai-basics-assessment-project-24"
version = "0.1.0"
description = "Color classifier for class assessment."
authors = ["Kashi_Grobe <kashigrobe@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pillow = "^10.3.0"
numpy = "^1.26.4"
opencv-python = "^4.9.0.80"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
